+++ skbase/utils/deep_equals/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+# copyright: skbase developers, BSD-3-Clause License (see LICENSE file)
+"""Module for nested equality checking."""
+from skbase.utils.deep_equals._deep_equals import deep_equals
+
+__all__ = [
+    "deep_equals",
+]
